requirements.txt
Django
Jinja2
python-frontmatter